requirements.txt
pandas
mlflow
notebook
scikit-learn
matplotlib
python-box
pyYAML
tqdm
ensure
joblib
types-pyYAML
Flask
Flask-Cors
psycopg2-binary

